apps/api/pyproject.toml
[project]
name = "company-platform-api"
version = "0.1.0"
description = "FastAPI backend for the company platform"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.14.0",
  "argon2-cffi>=23.1.0",
  "celery>=5.4.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "passlib>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.1",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.12",
  "redis>=5.2.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

